+++ equitania/res_users.py
@@ -26,32 +26,32 @@ from openerp import SUPERUSER_ID
 class eq_res_users(osv.osv):
     _inherit = 'res.users'
 
-    def _get_group(self,cr, uid, context=None):
-        dataobj = self.pool.get('ir.model.data')
-        result = super(eq_res_users, self)._get_group(cr, uid, context=context)
-        try:
-            dummy,group_purchase_id = dataobj.get_object_reference(cr, SUPERUSER_ID, 'equitania', 'purchase_in_products')
-            dummy,group_supplier_id = dataobj.get_object_reference(cr, SUPERUSER_ID, 'equitania', 'supplier_in_account')
-            dummy,group_access_reporting_menu_id = dataobj.get_object_reference(cr, SUPERUSER_ID, 'equitania', 'group_access_reporting')
-            dummy,group_access_employee_menu_id = dataobj.get_object_reference(cr, SUPERUSER_ID, 'equitania', 'group_access_hr_menu')
-            result.append(group_purchase_id)
-            result.append(group_supplier_id)
-            
-            result.append(group_access_reporting_menu_id)
-            result.append(group_access_employee_menu_id)
-        except ValueError:
-            # If these groups does not exists anymore
-            pass
-        return result
+    # def _get_group(self,cr, uid, context=None):
+    #     dataobj = self.pool.get('ir.model.data')
+    #     result = super(eq_res_users, self)._get_group(cr, uid, context=context)
+    #     try:
+    #         dummy,group_purchase_id = dataobj.get_object_reference(cr, SUPERUSER_ID, 'equitania', 'purchase_in_products')
+    #         dummy,group_supplier_id = dataobj.get_object_reference(cr, SUPERUSER_ID, 'equitania', 'supplier_in_account')
+    #         dummy,group_access_reporting_menu_id = dataobj.get_object_reference(cr, SUPERUSER_ID, 'equitania', 'group_access_reporting')
+    #         dummy,group_access_employee_menu_id = dataobj.get_object_reference(cr, SUPERUSER_ID, 'equitania', 'group_access_hr_menu')
+    #         result.append(group_purchase_id)
+    #         result.append(group_supplier_id)
+    #
+    #         result.append(group_access_reporting_menu_id)
+    #         result.append(group_access_employee_menu_id)
+    #     except ValueError:
+    #         # If these groups does not exists anymore
+    #         pass
+    #     return result
     
     _columns = {
                 'eq_employee_id': fields.many2one('hr.employee', 'Employee', copy=False),
                 'eq_signature' : fields.binary('Signature'),
                 }
 
-    _defaults = {
-        'groups_id': _get_group,
-    }
+    # _defaults = {
+    #     'groups_id': _get_group,
+    # }
     
     def write(self, cr, uid, ids, values, context={}):
         if context == None: